requirements.txt
fastapi==0.100.0
pydantic==2.0.2
pydantic_core==2.1.2
PyJWT==2.7.0
typing_extensions==4.7.1
uvicorn==0.22.0
